portfolio_manager/requirements.txt
# Portfolio Manager v8.1 - Requirements

# Core
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Data Fetching
yfinance>=0.2.30
ccxt>=4.0.0
fredapi>=0.5.0
MetaTrader5>=5.0.45

# Technical Analysis
ta>=0.11.0
pandas-ta>=0.3.14

# Machine Learning (for Hurst, etc.)
scikit-learn>=1.3.0

# API Server
fastapi>=0.104.0
uvicorn>=0.24.0

# Utilities
loguru>=0.7.0
pydantic>=2.5.0
python-dotenv>=1.0.0
schedule>=1.2.0
